OpCode Shell & IO v2
Revampped IO again, to take in a 'start address', and repeat it to the
user. The program also inserts dummy opcodes there (namely, Move.B). The
program identifies which opcodes were added (between Move.B, .W, and .L)
and prints out the opCode in a basic version  of the end IO message.