Implemented start is less than end check, and check line count.
To see the counter in action I recommend using addresses that are 40
apart, like 8000 for start 8040 for end. There is a bug where it only
prints out the press enter message once, but the user still needs to hit
enter to continue. I suggest making the addresses 80 apart to replicate
the bug.