Bcc OpCodes (ONLY BCC, BGE, BLT)
Only BCC, BGE, and BLT were listed as Bcc OpCodes that were required, so
those were the only ones that I did. These are complete for all sizes,
but I wasn't able to test them for 32-bit branches, because I couldn't
compile to that size. BSR is used substantially in the Demo file, and is
not complete here, so I'm not sure if it's something we need to do..