Fixed: CMP Dn, LSd (A#)
Fixed CMP OpCode for Dn: When Dn was 4 or more, it would drop the 4--
this was an ANDI issue (3, rather than 7) and was fixed. Verified this
issue wasn't copy-pasted anywhere else.
LSd and it's counterparts have an (A#) addressing mode that has no size
bits. Turns out the size is always .W, so I've added this and fixed the
spacing issue.